permissions

+++ app/Http/Controllers/ActivityLevelController.php
@@ -23,6 +23,9 @@ class ActivityLevelController extends Controller
     public function __construct()
     {
         $this->middleware('auth');
+        if(!$request->user()->can('edit-settings')){
+            return response(['Unauthorized(permissions)'], 401);
+        }
     }
 
     public function index(){
@@ -31,6 +34,9 @@ class ActivityLevelController extends Controller
     }
     
     public function create(Request $request){
+        if(!$request->user()->can('edit-settings')){
+            return response(['Unauthorized(permissions)'], 401);
+        }
         $this->validate($request, $this->validation);
         $this->removeConflict($request);
         $item = ActivityLevel::create($request->input());
@@ -43,6 +49,9 @@ class ActivityLevelController extends Controller
     }
     
     public function update($id, Request $request){
+        if(!$request->user()->can('edit-settings')){
+            return response(['Unauthorized(permissions)'], 401);
+        }
         $this->validate($request, $this->validation);
         $this->removeConflict($request);
         $item = ActivityLevel::findOrFail($id);
@@ -51,6 +60,9 @@ class ActivityLevelController extends Controller
     }
     
     public function delete($id){
+        if(!$request->user()->can('edit-settings')){
+            return response(['Unauthorized(permissions)'], 401);
+        }
         $item = ActivityLevel::findOrFail($id);
         $item->delete();
         return response([], 204);
