+++ app/Http/Controllers/OrderStatusTypeController.php
@@ -24,6 +24,9 @@ class OrderStatusTypeController extends Controller
     }
     
     public function create(Request $request){
+        if(!$request->user()->can('edit-settings')){
+            return response(['Unauthorized(permissions)'], 401);
+        }
         $validation = [
             'name' => 'string|required|min:1|max:255',
             'notes' => 'string|max:255'
@@ -39,6 +42,9 @@ class OrderStatusTypeController extends Controller
     }
     
     public function update($id, Request $request){
+        if(!$request->user()->can('edit-settings')){
+            return response(['Unauthorized(permissions)'], 401);
+        }
         $validation = [
             'name' => 'string|min:1|max:255|nullable',
             'notes' => 'string|max:255|nullable',
@@ -61,6 +67,9 @@ class OrderStatusTypeController extends Controller
     }
     
     public function delete($id){
+        if(!$request->user()->can('edit-settings')){
+            return response(['Unauthorized(permissions)'], 401);
+        }
         $item = OrderStatusType::findOrFail($id);
         $item->delete();
         return response([], 204);
